+++ src/Sir.Store/DocWriter.cs
@@ -16,15 +16,16 @@ namespace Sir.Store
             _stream = stream;
         }
 
-        public (long offset, int length) Append(IList<(uint keyId, uint valId)> doc)
+        public (long offset, int length) Append(IList<(long keyId, long valId)> doc)
         {
             var off = _stream.Position;
             var len = 0;
+
             foreach (var kv in doc)
             {
-                _stream.Write(BitConverter.GetBytes(kv.keyId), 0, sizeof(uint));
-                _stream.Write(BitConverter.GetBytes(kv.valId), 0, sizeof(uint));
-                len += sizeof(uint) * 2;
+                _stream.Write(BitConverter.GetBytes(kv.keyId), 0, sizeof(long));
+                _stream.Write(BitConverter.GetBytes(kv.valId), 0, sizeof(long));
+                len += sizeof(long) * 2;
             }
             
             return (off, len);
